# Detect exact device ID matches quickly