## Description

When a command is run with `--device` all relevant platform device discoverers gather all devices, which are then checked if the identifier or name exactly matches the specified device ID, then if they match the prefix of the specified device ID.  Some device discoverers with hard-coded return very quickly (`flutter-tester`, `chrome`, `macos`, `linux`), and others that shell out to underlying tools (Android, iOS, Fuschia) take much longer.

Instead of gathering all devices first, process them as they are discovered, and return quickly if an exact match is found.

Here are discovery times for exact matches on top of tree vs this PR (note these times include the overhead of calling a command I hacked the timing into, it's not how long this function took):
|   |master | PR   |
|---|---|---|
|flutter-tester   |1.49s |0.58s   |
|chrome   |1.51s |0.61s   |
|macos   |1.48s |0.59s   |
|iOS device identifier   |1.48s |1.27s   |

Before this change, all times were about the same, because it waited for all platform discoveries to be complete before matching.  After this change, the time takes as long as the relevant discoverer can return.

## Related Issues

Fixes https://github.com/flutter/flutter/issues/61466
Fixes https://github.com/flutter/flutter/issues/44898
Relevant for https://github.com/flutter/flutter/issues/15072 increasing discovery times when an exact networking iOS ID is specified without regressing other platforms.

## Tests
`getDeviceById exact matcher` uses new `LongPollingDeviceDiscovery` that never finishes polling. Exact matches are still returned quickly. This test runs forever on master.

## Checklist
- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I signed the [CLA].
- [x] I read and followed the [Flutter Style Guide], including [Features we expect every widget to implement].
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] All existing and new tests are passing.
- [x] The analyzer (`flutter analyze --flutter-repo`) does not report any problems on my PR.
- [x] I am willing to follow-up on review comments in a timely manner.

## Breaking Change
- [x] No, no existing tests failed, so this is *not* a breaking change.
- [ ] Yes, this is a breaking change. *If not, delete the remainder of this section.*