@jonahwilliams 2020-07-22
Very nice! I think you could construct a new test case that will now pass that would have failed. Consider two device discoverers, one returns instantly and one never returns.